+++ prisma/migrations/20260409145854_project_assignment_sync_role_from_float/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "ProjectAssignment" ADD COLUMN     "syncRoleFromFloat" BOOLEAN NOT NULL DEFAULT true;